feat: 使用float代替double (害)
Former-commit-id: acbf688ef5e6a373c8d4db1d4c9a87fbc5c33d3c

+++ Allay-API/src/main/java/cn/allay/api/entity/component/impl/base/EntityPlayerBaseComponent.java
@@ -45,8 +45,8 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent {
     void setHandSlot(@Range(from = 0, to = 8) int handSlot);
 
     @Override
-    default double getBaseOffset() {
-        return 1.62;
+    default float getBaseOffset() {
+        return 1.62f;
     }
 
     @Override
